+++ projects/mk-kit/src/lib/components/dialog/index.ts
@@ -1,3 +1,4 @@
 export * from './dialog';
 export * from './confirm-dialog';
+export * from './prompt-dialog';
 export * from './dialog.service';